pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "artist-preview"
version = "0.1.0"
description = "漫画/插图作品按画师分类工具"
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyperclip>=1.8.0",
    "pyyaml>=6.0",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/ArtistPreview"
"Bug Tracker" = "https://github.com/yourusername/ArtistPreview/issues"

[project.scripts]
artist-preview = "src.artist_preview:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
